Add LiveEventDetails component
