+++ frontend/sozisel-app/src/components/PresenterSession/EventsTimline/EventsTimeline.scss
@@ -1,9 +1,5 @@
 @import "../../../styles/styles.scss";
 
-.Timer {
-  text-align: center;
-}
-
 .EventsTimeline {
   height: 90%;
   .header {
