+++ frontend/sozisel-app/src/components/PresenterSession/EventsTimline/LiveEventDetails/LiveEventDetails.scss
@@ -1,0 +1,9 @@
+@import "../../../../styles/styles.scss";
+
+.Timer {
+  text-align: center;
+}
+
+.LiveEventsDetails {
+  margin: $spacing-02 $spacing-05;
+}
