+++ frontend/sozisel-app/src/components/TemplateCreation/Modules/Quiz/QuestionsList/QuestionsList.scss
@@ -1,6 +1,7 @@
 @import "../../../../../styles/styles";
 
 .QuestionsList {
+    margin-top: $spacing-04;
     text-align: center;
     background-color: $bg-03;
     padding: 0 $spacing-03;
